requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0
django-allauth==0.59.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
isodate==0.6.1
kombu==5.3.4
lxml==4.9.3
lxml==4.9.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
whitenoise==6.6.0
python-decouple==3.8
dj-rest-auth==5.0.2
Pillow==10.1.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
redis==5.0.1
celery==5.3.6